% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survode_control.R
\name{survode_control}
\alias{survode_control}
\title{Ancillary arguments for controlling survode fitting}
\usage{
survode_control(eps = 1e-04, maxit = 100)
}
\arguments{
\item{eps}{the convergence criteria for the optimization algorithm.
Default is 1e-4.}

\item{maxit}{the maximum number of iterations for the optimization algorithm.
Default is 100.}
}
\description{
Ancillary arguments for controlling survode fitting
}
